extract psalm commands into a library
