+++ lib/psalm/psalm.dart
@@ -1,0 +1,4 @@
+library psalm;
+
+export 'src/generate_config_file.dart';
+export 'src/run.dart';
