+++ lib/psalm/src/run.dart
@@ -1,9 +1,6 @@
 import 'dart:io';
 
-import 'package:path/path.dart' as p;
-
-Future<int> run(
-    Directory projectLocation, String psalmConfigLocation) async {
+Future<int> run(Directory projectLocation, String psalmConfigLocation) async {
   var result = await Process.run(
       './vendor-bin/errors_over_time/vendor/vimeo/psalm/psalm',
       [
@@ -22,24 +19,6 @@ Future<int> run(
     throw Exception(
         'psalm returned the following exit code ${result.exitCode} with stderr ${result.stderr}');
   }
-}
-
-Future<File> generateConfigurationFile(Directory projectLocation) async {
-    var result = await Process.run(
-      './vendor-bin/errors_over_time/vendor/vimeo/psalm/psalm',
-      [
-        '--init',
-      ],
-      workingDirectory: projectLocation.path);
-  if (result.exitCode != 0) {
-    throw Exception(
-        'psalm init returned the following exit code ${result.exitCode} with stderr ${result.stderr}');
-  }
-  var psalmConfig = File(p.join(projectLocation.path, 'psalm.xml'));
-  if (!(await psalmConfig.exists())){
-    throw Exception('Failed to create psalm config file');
-  }
-  return psalmConfig;
 }
 
 int _numberOfErrosFromPsalmOutput(String psalmOutput) {
